requirements.txt
# Data handling
pandas
numpy

# Environment variables
python-dotenv

# Configuration
pyyaml

# Logging (optional if using custom logger)
loguru==0.7.0

# Visualization (for later analysis)
matplotlib
seaborn


# Utilities
requests==2.31.0   # For API fetching if needed
pyarrow     # For Parquet file handling
pyspark

# Testing
pytest
pytest-cov
